requirements.txt
streamlit
cohere
fpdf2
python-dotenv
whisper @ git+https://github.com/openai/whisper.git
gTTS
soundfile
numpy
streamlit-webrtc





